+++ backend/Dockerfile
@@ -21,4 +21,3 @@ COPY . .
 ENV PORT=$PDT_API_PORT
 
 CMD /wait && ./run_dev.sh
-
