+++ backend/api.py
@@ -11,6 +11,7 @@ from backend.auth import user_manager, jwt, refresh_token
 from backend.schemas import spec
 from backend.routes.partners import bp as partners_bp
 from backend.routes.incidents import bp as incidents_bp
+from backend.routes.officers import bp as officers_bp 
 from backend.routes.auth import bp as auth_bp
 from backend.routes.healthcheck import bp as healthcheck_bp
 from backend.utils import dev_only
@@ -130,6 +131,7 @@ def register_routes(app: Flask):
     app.register_blueprint(incidents_bp)
     app.register_blueprint(auth_bp)
     app.register_blueprint(healthcheck_bp)
+    app.register_blueprint(officers_bp)
 
     @app.route("/")
     def hello_world():
