+++ frontend/compositions/search-panel/search-panel.tsx
@@ -47,6 +47,7 @@ export const SearchPanel = (props: SearchPanelProps) => {
     } catch (e) {
       console.error("Unexpected search error", e)
       setErrorMessage("Something went wrong. Please try again.")
+      
       /* # TODO: Add error handling when a 401 is recieved. Redirect to login */
     }
     setIsLoading(false)
